server/requirements.txt
fastapi
uvicorn[standard]
google-generativeai
python-dotenv
python-multipart
pillow
pytest
httpx
supabase>=2.0
pyjwt[crypto]>=2.8
